pyproject.toml
[project]
name = "findresearchpapers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "datasets>=4.1.1",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "langchain-postgres>=0.0.15",
    "notebook>=7.4.5",
    "pre-commit>=4.3.0",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pyright>=1.1.405",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.1",
    "uvicorn>=0.37.0",
]
